+++ src/Type/SchemaType.php
@@ -150,11 +150,11 @@ final class SchemaType extends AbstractType implements DatabaseTypeInterface, \S
         foreach ($this->getSchema()->all() as $element) {
             $data[$element->name()] = null;
             if (\array_key_exists($element->name(), $value) && $value[$element->name()] !== null) {
-                $data[$element->name()] = Type::create($value[$element->name()], $element->type());
-            }
-
-            if ($element instanceof TransformableInterface) {
-                $data[$element->name()] = $element->transform($data[$element->name()]);
+                if ($element instanceof TransformableInterface) {
+                    $data[$element->name()] = $element->transform($value[$element->name()]);
+                } else {
+                    $data[$element->name()] = Type::create($value[$element->name()], $element->type());
+                }
             }
         }
 
@@ -207,7 +207,7 @@ final class SchemaType extends AbstractType implements DatabaseTypeInterface, \S
             return $this->schema;
         }
 
-        throw new \Exception("Cant initialize without schema");
+        throw new \Exception('Cant initialize without schema');
     }
 
     public function __get($name)
@@ -291,8 +291,8 @@ final class SchemaType extends AbstractType implements DatabaseTypeInterface, \S
     {
         /** @var SchemaType $schemaType */
         $schemaType = Type::get(SchemaType::serviceName());
-        $this->builder = $schemaType->builder;
         $this->serviceManager = $schemaType->serviceManager;
+        $this->builder = $schemaType->builder;
 
         parent::unserialize($serialized);
     }